pyproject.toml
[project]
name = "lifersr"
version = "0.1.0"
description = "LifeRSR — Real to Simulation to Real. Agent-Native Personal Knowledge Base (AI 参谋知识库)"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "python-frontmatter>=1.1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
]

[project.scripts]
kb = "src.cli:cli"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
